+++ src/file_spec.rs
@@ -1,8 +1,7 @@
-use crate::now_local;
-use crate::FlexiLoggerError;
+use crate::{DeferredNow, FlexiLoggerError};
 use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
-use time::{format_description::FormatItem, macros::format_description, OffsetDateTime};
+use time::{format_description::FormatItem, macros::format_description};
 
 /// Builder object for specifying the name and path of the log output file.
 ///
@@ -184,7 +183,7 @@ impl FileSpec {
             filename.push('_');
             filename.push_str(discriminant);
         }
-        if let Some(timestamp) = &self.timestamp_cfg.get_timestamp(self.use_utc) {
+        if let Some(timestamp) = &self.timestamp_cfg.get_timestamp() {
             filename.push_str(timestamp);
         }
         if let Some(infix) = o_infix {
@@ -209,7 +208,7 @@ impl FileSpec {
             filename.push('_');
             filename.push_str(discriminant);
         }
-        if let Some(timestamp) = &self.timestamp_cfg.get_timestamp(self.use_utc) {
+        if let Some(timestamp) = &self.timestamp_cfg.get_timestamp() {
             filename.push_str(timestamp);
         }
         if let Some(infix) = o_infix {
@@ -244,11 +243,11 @@ enum TimestampCfg {
     No,
 }
 impl TimestampCfg {
-    fn get_timestamp(&self, use_utc: bool) -> Option<String> {
+    fn get_timestamp(&self) -> Option<String> {
         match self {
             Self::Default | Self::Yes => {
                 Some(
-                    if use_utc {OffsetDateTime::now_utc()} else {now_local()}
+                    DeferredNow::now_local()
                     .format(TS_USCORE_DASHES_USCORE_DASHES).unwrap(/*ok*/),
                 )
             }